+++ templates/common/packages/config/endpoints.js
@@ -4,7 +4,9 @@ export const endpoints = {
     home: 'home/1'
   },
   apiGee: {
-    todo: 'todos/1'
+    todo: 'todos/1',
+    users: 'profile',
+    account: 'account',
   }
 };
 
@@ -14,12 +16,26 @@ export const endpoints = {
  */
 export const contentDomain = 'https://jsonplaceholder.typicode.com/';
 export const apiGeeDomain = 'https://jsonplaceholder.typicode.com/';
+export const apiGeeDomainMockRest = 'http://localhost:5002/';
+export const apiGeeDomainMockGraphQL = 'http://localhost:5004/';
 
-export const mockRestDomain = 'http://localhost:5002/';
-export const mockGraphQLDomain = 'http://localhost:5004/';
-export const apiDomain = 'http://qa.somedomain.com/';
+/** getContentServiceUrl - return content service url to be passed to fetch api
+ * (key, isMock) - > key - get url from endpoints ;  isMock - serve data from public folder
+ */
+ export const getContentServiceUrl = (key, isMock = false) => {
+  const url = endpoints.content[key];
+  return isMock ? `/static/mock/${url}.json` : `${contentDomain}${url}`;
+};
 
-export const apiEndPoints = {
-  users: 'profile',
-  account: 'account',
+/** getDataServiceUrl - return data service url to be passed to fetch api
+ * (key, isMock) - > key - get url from endpoints ;  isMock - serve data from public folder
+ */
+export const getDataServiceUrl = (key, isMock = false) => {
+  const url = endpoints.apiGee[key];
+  return sessionStorage.getItem('serveMock') === 'true' || isMock ? `${apiGeeDomainMockRest}${url}` : `${apiGeeDomain}${url}`;
+};
+
+export const getGraphQlServiceUrl = (key, isMock = false) => {
+  const url = endpoints.apiGee[key];
+  return sessionStorage.getItem('serveMock') === 'true' || isMock ? `${apiGeeDomainMockGraphQL}${url}` : `${apiGeeDomain}${url}`;
 };
